escape_spawn_behavior

+++ src/systems/game/components.rs
@@ -2,15 +2,15 @@ use crate::prelude::*;
 
 #[derive(Component)]
 pub struct Player {
-    pub health: u32,
-    pub mana: u32,
-    pub damage: u32,
+    pub health: f32,
+    pub mana: f32,
+    pub damage: f32,
     pub mov_speed: f32,
-    pub speed: u32
+    pub speed: f32
 }
 
 impl Default for Player {
     fn default() -> Self {
-        Self { health: 100, mana: 100, damage: 10, mov_speed: 500.0, speed: 100 }
+        Self { health: 100., mana: 100., damage: 10., mov_speed: 500.0, speed: 100. }
     }
 }
