+++ src/systems/game/fight/turns/events.rs
@@ -1,7 +1,7 @@
 use bevy::prelude::*;
 
 #[derive(Event)]
-pub struct PlayerDamageEvent(pub u32);
+pub struct PlayerDamageEvent(pub f32);
 
 #[derive(Event)]
-pub struct EnemyDamageEvent(pub u32);
+pub struct EnemyDamageEvent(pub f32);
